Add Fleet management features including DTOs, controller methods, and service logic; implement fleet creation and retrieval by owner

+++ src/main/java/com/hush0k/pirateTeam/ship/dto/request/ShipUpdateDto.java
@@ -21,7 +21,7 @@ public record ShipUpdateDto(
         Integer maxCrew,
 
         @Positive(message = "Max speed must be positive")
-        Float maxSpeed,
+        Integer maxSpeed,
 
         Country builderCountry,
